Added SOAP Fault and Unit Tests
Added SOAP Fault and Endpoint Unit testing

+++ spring-boot-restapi-consume-soapservice/soap-ws-jpa-crud/src/main/java/com/vision/tech/soap/service/EmployeeService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.vision.tech.soap.entities.Employee;
+import com.vision.tech.soap.exception.EmployeeNotFoundException;
 import com.vision.tech.soap.repo.EmployeeRepository;
 
 @Service
@@ -16,9 +17,11 @@ public class EmployeeService implements IEmployeeService {
 
 	@Override
 	public Employee getEmployeeById(long employeeId) {
-
-		Employee obj = employeeRepository.findByEmployeeId(employeeId);
-		return obj;
+		Employee employee = employeeRepository.findByEmployeeId(employeeId);
+		if (employee == null) {
+			throw new EmployeeNotFoundException("Invalid employee Id:" +employeeId);
+		}
+		return employee;
 
 	}
 
@@ -29,14 +32,21 @@ public class EmployeeService implements IEmployeeService {
 
 	@Override
 	public void updateEmployee(Employee employee) {
+		Employee emp = employeeRepository.findByEmployeeId(employee.getEmployeeId());
+		if (emp == null) {
+			throw new EmployeeNotFoundException("Employee not present");
+		}
 		employeeRepository.save(employee);
 
 	}
 
 	@Override
 	public void deleteEmployee(long employeeId) {
+		Employee employee = employeeRepository.findByEmployeeId(employeeId);
+		if (employee == null) {
+			throw new EmployeeNotFoundException("Invalid employee Id: " +employeeId);
+		}
 		employeeRepository.deleteById(employeeId);
-
 	}
 
 	@Override
